requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
python-vlc>=3.0.20000